+++ lib/Exception/WebDriverException.php
@@ -5,9 +5,12 @@ namespace Facebook\WebDriver\Exception;
 use Exception;
 
 /**
+ * Ancestor for all exceptions defined in W3C WebDriver standard.
+ * (And also for deprecated JsonWire protocol exceptions.)
+ *
  * @see https://w3c.github.io/webdriver/#errors
  */
-class WebDriverException extends Exception
+class WebDriverException extends Exception implements PhpWebDriverExceptionInterface
 {
     private $results;